+++ gsjson.js
@@ -10,8 +10,6 @@ program
     .version(packageData.version)
     .usage('<spreadsheet-id> [file] [options]')
     .option('-b, --beautify', 'Beautify final JSON')
-    .option('-u, --user <user>', 'User to login')
-    .option('-p, --password <password>', 'Password to login')
     .option('-t, --token <token>', 'Auth token acquired externally')
     .option('-y, --tokentype <tokentype>', 'Type of the informed token (defaults to Bearer)')
     .option('-w, --worksheet <n>', 'Worksheet index or title (defaults to first worksheet, can be repeated)', handleWorksheetIdentifiers)